sale_mrp_bom_configurable/test-requirements.txt
odoo-addon-mrp_bom_configurable @ git+https://github.com/OCA/manufacture.git@refs/pull/1079/head#subdirectory=setup/mrp_bom_configurable
